pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "bayer-visualiser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.8.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tabulate>=0.9.0",
    "watchdog>=6.0.0",
]
